Move and rename StringPtr in module
The already implemented function that can read a
string s now in a separate string

We also want to be able to write the poiter in the linear mem
and so we add a malloc function that has to be implemented
